pyproject.toml
[tool.poetry]
name = "hume"
version = "0.1.0"
description = ""
authors = ["Rashicom <rashid.kp484@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
django-environ = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
